Split Invoice.cs up into parts
Created a separate enum file and there was already a save method in the repository so I also removed this. Once this was removed there was no need or use for injecting the InvoiceRepository. Once I removed this there were issues with the constructor creating Invoice objects in the test cases so I updated them to not use any param